Adv Gridworld Integration
I forgot to make a copy of QLearning, so I need to revert it before
pushing this to master.

Comparison_advgridworld.py will attempt to run the three algorithms and
print the graph, but right now it's printing a straight line for shared
autonomy and nothing for the other two.